refactor(#297): extract queue management routes to routers/queue.py
Move queue-related HTTP endpoints and logic from server.py into a dedicated
queue router module. This includes:

- GET /api/queue — aggregated queue status across 15 active repos
- POST /api/runs/{repo}/cancel/{run_id} — cancel single workflow run
- POST /api/runs/{repo}/rerun/{run_id} — re-run failed jobs
- POST /api/queue/cancel-workflow — bulk cancel runs by workflow name
- GET /api/queue/diagnose — detailed bottleneck analysis with runner pool diagnostics

Helper functions (_queue_impl, _empty_queue_result, _get_recent_org_repos,
_fetch_repo_runs) are moved to the queue router to keep queue logic isolated.

Also add cache_delete() utility to cache_utils for explicit cache invalidation
(used for clearing queue/diagnose caches after mutations).

The extraction preserves all behavior and maintains internal dependencies:
- _queue_impl remains callable from server.py stats aggregation
- Rate limiting and safety checks unchanged
- Response schemas identical
- All diagnostics logic (pick-runner detection, group restrictions, labels)
  fully preserved

Tests still passing; all queue endpoints verified functional.

Follows the established pattern from batch-2 router extractions (fleet, runners,
runs_workflows, system).
